+++ R/ard_emmeans_emmeans.R
@@ -1,57 +1,44 @@
-#' ARD for LS Mean Difference
-#'
 #' @description
-#' This function calculates least-squares mean differences using the 'emmeans'
+#' The `ard_emmeans_emmeans()` function calculates least-squares means using the 'emmeans'
 #' package using the following
 #'
 #' ```r
 #' emmeans::emmeans(object = <regression model>, specs = ~ <primary covariate>) |>
-#'   emmeans::contrast(method = "pairwise") |>
-#'   summary(infer = TRUE, level = <confidence level>)
+#'   summary(emmeans, calc = c(n = ".wgt."))
 #' ```
 #'
 #' The arguments `data`, `formula`, `method`, `method.args`, `package` are used
 #' to construct the regression model via `cardx::construct_model()`.
 #'
-#' @param data (`data.frame`/`survey.design`)\cr
-#'   a data frame or survey design object
-#' @inheritParams construct_model
-#' @param response_type (`string`)
-#'   string indicating whether the model outcome is `'continuous'`
-#'   or `'dichotomous'`. When `'dichotomous'`, the call to `emmeans::emmeans()` is
-#'   supplemented with argument `regrid="response"`.
-#' @param conf.level (scalar `numeric`)\cr
-#'   confidence level for confidence interval. Default is `0.95`.
-#' @param primary_covariate (`string`)\cr
-#'   string indicating the primary covariate (typically the dichotomous treatment variable).
-#'   Default is the first covariate listed in the formula.
-#'
-#' @return ARD data frame
 #' @export
+#' @rdname ard_emmeans
 #'
 #' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "emmeans"))
-#' ard_emmeans_mean_difference(
+#' # LS Means
+#' ard_emmeans_emmeans(
 #'   data = mtcars,
 #'   formula = mpg ~ am + cyl,
 #'   method = "lm"
 #' )
 #'
-#' ard_emmeans_mean_difference(
+#' ard_emmeans_emmeans(
 #'   data = mtcars,
 #'   formula = vs ~ am + mpg,
 #'   method = "glm",
 #'   method.args = list(family = binomial),
 #'   response_type = "dichotomous"
 #' )
-ard_emmeans_mean_difference <- function(data, formula, method,
-                                        method.args = list(),
-                                        package = "base",
-                                        response_type = c("continuous", "dichotomous"),
-                                        conf.level = 0.95,
-                                        primary_covariate =
-                                          stats::terms(formula) |>
-                                            attr("term.labels") |>
-                                            getElement(1L)) {
+ard_emmeans_emmeans <- function(data,
+                                formula,
+                                method,
+                                method.args = list(),
+                                package = "base",
+                                response_type = c("continuous", "dichotomous"),
+                                conf.level = 0.95,
+                                primary_covariate =
+                                  stats::terms(formula) |>
+                                    attr("term.labels") |>
+                                    getElement(1L)) {
   set_cli_abort_call()
 
   # check package installation -------------------------------------------------
@@ -70,45 +57,56 @@ ard_emmeans_mean_difference <- function(data, formula, method,
   data_in <- if (dplyr::last(class(data)) == "survey.design") data$variables else data
 
   # build ARD ------------------------------------------------------------------
-  cards::ard_complex(
+  result <- cards::ard_mvsummary(
     data = data_in,
     variables = all_of(primary_covariate),
     statistic = all_of(primary_covariate) ~ list(
       emmeans =
-        .calc_emmeans_mean_difference(
-          data, formula, method, {{ method.args }}, package, response_type, conf.level, primary_covariate
+        .calc_emmeans(
+          data = data, formula = formula, method = method,
+          method.args = {{ method.args }}, package = package,
+          response_type = response_type, conf.level = conf.level,
+          primary_covariate = primary_covariate
         )
     )
-  ) |>
+  )
+  # unlist stat column
+  if (length(result$stat[[which(result$stat_label == "variable_level")]]) > 1) {
+    result <- result |> tidyr::unnest_longer(col = "stat")
+  }
+
+  result |>
     dplyr::select(-"stat_label") |>
     dplyr::left_join(
-      .df_ttest_stat_labels(primary_covariate),
+      .df_emmeans_stat_labels("emmeans"),
       by = "stat_name"
     ) |>
     dplyr::mutate(
       variable = "contrast",
       variable_level = if ("variable_level" %in% .data$stat_name) {
-        .data$stat[.data$stat_name == "variable_level"]
+        rep_len(.data$stat[.data$stat_name == "variable_level"], length.out = nrow(result))
       } else {
         NA
       },
       group1 = .env$primary_covariate,
       stat_label = dplyr::coalesce(.data$stat_label, .data$stat_name),
-      context = "emmeans_mean_difference",
+      context = "emmeans_emmeans"
     ) |>
+    dplyr::filter(!is.na(.data$stat)) |>
     dplyr::filter(.data$stat_name != "variable_level") |>
+    dplyr::arrange(.data$variable_level) |>
     cards::as_card() |>
     cards::tidy_ard_column_order() |>
     cards::tidy_ard_row_order()
 }
 
-# function to perform calculations -------------------------------------------
-.calc_emmeans_mean_difference <- function(data, formula, method,
-                                          method.args,
-                                          package,
-                                          response_type,
-                                          conf.level,
-                                          primary_covariate) {
+# function to perform calculations ---------------------------------------------
+.calc_emmeans <- function(data, formula, method,
+                          method.args,
+                          package,
+                          response_type,
+                          conf.level,
+                          primary_covariate) {
   cards::as_cards_fn(
     \(x, ...) {
       # construct primary model ------------------------------------------------
@@ -128,36 +126,38 @@ ard_emmeans_mean_difference <- function(data, formula, method,
           code = do.call("emmeans", args = emmeans_args)
         )
 
-      df_results <-
-        emmeans |>
-        emmeans::contrast(method = "pairwise") |>
-        summary(infer = TRUE, level = conf.level)
+      # calculate mean estimates ---------------------------------------------
+      results <-
+        summary(emmeans, calc = c(n = ".wgt.")) |>
+        dplyr::as_tibble() |>
+        dplyr::rename(
+          estimate = any_of(c("emmean", "prob")),
+          n = any_of("n")
+        ) |>
+        dplyr::rename(variable_level = all_of(primary_covariate)) |>
+        dplyr::mutate(variable_level = as.character(.data$variable_level))
 
       # convert results to ARD format ------------------------------------------
-      df_results |>
+      results |>
         dplyr::as_tibble() |>
         dplyr::rename(
           conf.low = any_of("asymp.LCL"),
           conf.high = any_of("asymp.UCL"),
           conf.low = any_of("lower.CL"),
-          conf.high = any_of("upper.CL")
-        ) %>%
-        dplyr::select(
-          variable_level = "contrast",
-          "estimate",
-          std.error = "SE", "df",
+          conf.high = any_of("upper.CL"),
+          std.error = any_of("SE")
+        ) |>
+        dplyr::select(any_of(c(
+          "variable_level", "estimate",
+          "std.error", "df", "n",
           "conf.low", "conf.high", "p.value"
-        ) %>%
+        ))) |>
         dplyr::mutate(
           conf.level = .env$conf.level,
-          method =
-            ifelse(
-              length(attr(stats::terms(formula), "term.labels") |> discard(~ startsWith(., "1 |"))) == 1L,
-              "Least-squares mean difference",
-              "Least-squares adjusted mean difference"
-            )
-        )
+          method = "Least-squares means"
+        ) |>
+        dplyr::mutate(across(everything(), ~ .x |> as.list()))
     },
-    stat_names = c("estimate", "std.error", "df", "conf.low", "conf.high", "p.value", "conf.level", "method")
+    stat_names = c("variable_level", "estimate", "std.error", "df", "conf.low", "conf.high", "p.value", "conf.level", "method", "n")
   )
 }