+++ tests/testthat/test-ard_emmeans_emmeans.R
@@ -1,97 +1,85 @@
-skip_if_not(is_pkg_installed(c("emmeans", "survey", "lme4")))
+skip_if_pkg_not_installed(c("emmeans", "survey", "lme4"))
 
-test_that("ard_emmeans_mean_difference() works", {
+test_that("ard_emmeans_emmeans() works", {
   withr::local_options(width = 250)
 
-  expect_error(
-    ard_emmeans_mean_difference <-
-      ard_emmeans_mean_difference(
+  expect_silent(
+    ard_emmeans_emmeans <-
+      ard_emmeans_emmeans(
         data = mtcars,
         formula = vs ~ am + mpg,
         method = "glm",
         method.args = list(family = binomial),
         response_type = "dichotomous"
-      ),
-    NA
+      )
   )
-  expect_snapshot(ard_emmeans_mean_difference |> print(columns = "all"))
+  expect_snapshot(ard_emmeans_emmeans |> print(columns = "all"))
 
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "method"),
-    list(method = "Least-squares adjusted mean difference")
+    cards::get_ard_statistics(ard_emmeans_emmeans, stat_name %in% "method")[1],
+    list(method = "Least-squares means")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_emmeans, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     glm(vs ~ am + mpg, data = mtcars, family = binomial) |>
       emmeans::emmeans(specs = ~am, regrid = "response") |>
-      emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
-      getElement("estimate")
+      getElement("prob")
   )
 
-
-  expect_error(
-    ard_emmeans_mean_difference_lme4 <-
-      ard_emmeans_mean_difference(
+  expect_silent(
+    ard_emmeans_emmeans_lme4 <-
+      ard_emmeans_emmeans(
         data = mtcars,
         formula = vs ~ am + (1 | cyl),
         method = "glmer",
         method.args = list(family = binomial),
         package = "lme4",
         response_type = "dichotomous"
-      ),
-    NA
+      )
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "method"),
-    list(method = "Least-squares mean difference")
+    cards::get_ard_statistics(ard_emmeans_emmeans_lme4, stat_name %in% "method")[1],
+    list(method = "Least-squares means")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_emmeans_lme4, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     lme4::glmer(vs ~ am + (1 | cyl), data = mtcars, family = binomial) |>
       emmeans::emmeans(specs = ~am, regrid = "response") |>
-      emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
-      getElement("estimate")
+      getElement("prob")
   )
 
-
   #styler: off
-  expect_error({
+  expect_silent({
     data(api, package = "survey")
-    ard_emmeans_mean_difference_svy <-
+    ard_emmeans_emmeans_svy <-
       survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc) |>
-      ard_emmeans_mean_difference(
+      ard_emmeans_emmeans(
         formula = api00 ~ sch.wide,
         method = "svyglm",
         package = "survey"
-      )},
-    NA
+      )}
   )
   # styler: on
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "method"),
-    list(method = "Least-squares mean difference")
-  )
-  expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_emmeans_svy, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     survey::svyglm(api00 ~ sch.wide, design = survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)) |>
       emmeans::emmeans(specs = ~sch.wide, regrid = "response") |>
-      emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
-      getElement("estimate")
+      getElement("emmean")
   )
 })
 
-test_that("ard_emmeans_mean_difference() follows ard structure", {
+test_that("ard_emmeans_emmeans() follows ard structure", {
   expect_silent(
-    ard_emmeans_mean_difference(
+    ard_emmeans_emmeans(
       data = mtcars,
       formula = vs ~ am + mpg,
       method = "glm",
@@ -102,11 +90,11 @@ test_that("ard_emmeans_mean_difference() follows ard structure", {
   )
 })
 
-test_that("ard_emmeans_mean_difference() errors are returned correctly", {
+test_that("ard_emmeans_emmeans() errors are returned correctly", {
   withr::local_options(width = 250)
 
   expect_silent(
-    ard <- ard_emmeans_mean_difference(
+    ard <- ard_emmeans_emmeans(
       data = mtcars,
       formula = vs ~ am + mpg,
       method = "glm",