+++ NEWS.md
@@ -1,3 +1,11 @@
+# cardx 0.3.1
+
+* Added the `ard_emmeans_emmeans()` function. (#317)
+
+* Renamed `ard_emmeans_mean_difference()` to `ard_emmeans_contrast()` to align with function naming conventions.
+
+* The `ard_complex()` function has been renamed to `ard_mvsummary()`.
+
 # cardx 0.3.0
 
 ## New Features and Updates